+++ ionics_fits/models/containers.py
@@ -1,4 +1,3 @@
-import copy
 from typing import Dict, List, Optional, Tuple, TYPE_CHECKING
 
 import numpy as np
@@ -16,9 +15,11 @@ if TYPE_CHECKING:
 class AggregateModel(Model):
     """Model formed by aggregating one or more models.
 
+    When aggregating a number of identical models, use a :class RepeatedModel: instead.
+
     Aggregate models have a number of uses. For example:
       - joint fits to multiple data sets (pass the datasets in as y channels and use
-        "bound" parameters for any parameters which are fit jointly to all y channels).
+        "common" parameters for any parameters which are fit jointly to all y channels).
       - fit multiple datasets simultaneously. This is useful, for example in automated
       tooling such as `ndscan`'s `OnlineAnalysis`.
     """
@@ -32,11 +33,13 @@ class AggregateModel(Model):
     ):
         """
         :param models: The models to be aggregated. This should be a dictionary mapping
-          model names to model instances. The model names are used as prefixes for names
+          model names to model instances. The model names are used as suffixes for names
           of model parameters and derived results. For example, if one of the aggregated
-          models named `foo` has a parameter `bar`, the aggregate model will have a
-          parameter `bar_foo`. The passed-in models are considered "owned" by the
-          AggregateModel and should not be reused / modified externally.
+          models named `model` has a parameter `param`, the aggregate model will have a
+          parameter `param_model`.
+
+          The passed-in models are considered "owned" by the AggregateModel and should
+          not be used / modified elsewhere.
 
         :param common_params: Optional dictionary specifying "common" model parameters.
           This feature allows multiple parameters (which can be from the same or
@@ -51,15 +54,14 @@ class AggregateModel(Model):
           The dictionary keys are the names of the new model parameters.
 
           The dictionary values are a tuple containing the new model template parameter
-          and a list of parameters to bind to the new parameter.
+          and a list of parameters to bind to the new parameter. The bound parameter
+          lists should be lists of tuples, specifying the parameters to bind to the new
+          parameter. The tuples should contain two strings, specifying the name of the
+          model which owns the common parameter, and the name of the model parameter to
+          make common.
 
           The new model parameters inherit their metadata (limits etc.) from the
           template parameters, which are (deep) copied and are not modified.
-
-          The bound parameter lists should be lists of tuples, specifying the parameters
-          to bind to the new parameter. The tuples should contain two strings,
-          specifying the name of the model which owns the common parameter, and the name
-          of the model parameter to make common.
 
         At present this class only supports models with a single y channel. This is just
         because no one got around to implementing it yet rather than any fundamental
@@ -154,8 +156,8 @@ class AggregateModel(Model):
         x: Array[("num_samples",), np.float64],
         param_values: Dict[str, float],
     ) -> Array[("num_y_channels", "num_samples"), np.float64]:
-        ys = []
-        for model_name, model in self.__models.items():
+        ys = np.zeros((self.get_num_y_channels(), len(x)), dtype=np.float64)
+        for idx, (model_name, model) in enumerate(self.__models.items()):
             model_common_params = self.__model_common_params[model_name]
             model_params = {
                 param_name: param_values[f"{param_name}_{model_name}"]
@@ -171,9 +173,9 @@ class AggregateModel(Model):
                 }
             )
 
-            ys.append(model.func(x, model_params))
+            ys[idx, :] = np.atleast_2d(model.func(x, model_params))
 
-        return np.array(ys)
+        return ys
 
     def estimate_parameters(
         self,
@@ -191,8 +193,8 @@ class AggregateModel(Model):
 
             model.estimate_parameters(x, y[idx])
 
+        # use the mean value from all models as our heuristic for common params
         for new_param_name, binds in self.__common_param_list.items():
-
             estimates = [
                 self.__models[model_name].parameters[param_name].get_initial_value()
                 for model_name, param_name in binds
@@ -241,7 +243,7 @@ class AggregateModel(Model):
 
             derived = model.calculate_derived_params(
                 x=x,
-                y=y[idx],
+                y=y[idx, :],
                 fitted_params=model_fitted_params,
                 fit_uncertainties=model_fit_uncertainties,
             )
@@ -249,13 +251,13 @@ class AggregateModel(Model):
 
             derived_params.update(
                 {
-                    f"{model_name}_{param_name}": value
+                    f"{param_name}_{model_name}": value
                     for param_name, value in model_derived_params.items()
                 }
             )
             derived_uncertainties.update(
                 {
-                    f"{model_name}_{param_name}": value
+                    f"{param_name}_{model_name}": value
                     for param_name, value in model_derived_uncertainties.items()
                 }
             )
@@ -264,10 +266,10 @@ class AggregateModel(Model):
 
 
 class RepeatedModel(Model):
-    """Model formed by repeating a `Model` one more times
+    """Model formed by repeating a `Model` one more times.
 
     The `RepeatedModel` has multiple y-channels, corresponding to the repetitions of
-    the wrapped model.
+    the repeated model.
 
     Repeated models allow multiple datasets to be analysed simultaneously. This is
     useful, for example, when doing joint fits to datasets (using common parameters)
@@ -277,137 +279,159 @@ class RepeatedModel(Model):
 
     def __init__(
         self,
-        inner: Model,
+        model: Model,
         common_params: Optional[List[str]] = None,
         num_repetitions: int = 2,
+        aggregate_results=False,
     ):
         """
-        :param inner: The wrapped model, the implementation of `inner` will be used to
-          generate data for the y channels. This model is considered owned by the
+        :param model: The repeated model, the implementation of `model` will be used to
+          generate data for all y channels. This model is considered owned by the
           RepeatedModel and should not be used / modified elsewhere.
-        :param common_params: optional list of names of model arguments, whose value is
-          common to all y channels. All other model parameters are independent
-        :param num_repetitions: the number of times the inner model is repeated
+        :param common_params: optional list of names of model parameters, whose value is
+          common to all y channels. All other model parameters are allowed to vary
+          independently between the y channels
+        :param num_repetitions: the number of times the model is repeated
+        :param aggregate_results: determines whether derived results are aggregated or
+          not (see below). The default behaviour is to not aggregate results. This is
+          generally suitable when one wants access to the values of non-common
+          parameters from the various repetitions. Aggregating results can be useful,
+          for example, when all parameters are common across the repetitions and one
+          wants a single set of values reported.
 
-        Parameters of the new model:
-          - all common parameters of the inner model are parameters of the outer model
-          - for each independent (not common) or derived parameter of the inner model,
-            `foo`, the outer model has parameters `foo_{n}` for n in
+        The `RepeatedModel` has the following parameters and fit results:
+          - all common parameters of the repeated model are parameters of the new model
+          - for each independent (not common) parameter of the repeated model, `param`,
+            the `RepeatedModel` has parameters `param_{n}` for n in
             [0, .., num_repitions-1]
-          - for each independent or derived parameter, `foo`, the outer model calculates
-            statistics for the results from the various models: `foo_mean` and
-            `foo_peak_peak`.
+          - for each independent parameter, `param`, the `RepeatedModel` model
+            has additional fit results representing statistics between the repetitions.
+            For each independent parameter `param`, the results dictionary will have
+            additional quantities: `param_mean` and `param_peak_peak`.
+
+        If :param aggregate_results: is `False` the rules for fit results follow those
+        described previously. This is the default behaviour.
+
+        If `aggregate_results` is `True`:
+          - no statistical quantities are calculated for fit parameters or derived
+            results
+          - all derived results whose values and uncertainties are the same for all
+            repetitions are aggregated together to give a single result. This has the
+            same name as the original model (no suffix).
+          - all derived results whose value is not the same for al repetitions are
+            omitted.
         """
-        inner_params = set(inner.parameters.keys())
+        self.__aggregate_results = aggregate_results
+        model_params = set(model.parameters.keys())
         common_params = set(common_params or [])
 
-        if not common_params.issubset(inner_params):
+        if not common_params.issubset(model_params):
             raise ValueError(
-                "Common parameters must be a subset of the inner model's parameters"
+                "Common parameters must be a subset of the model's parameters"
             )
 
-        independent_params = set(inner.parameters.keys()) - common_params
-        params = {param: inner.parameters[param] for param in common_params}
+        independent_params = set(model_params) - common_params
+        params = {param: model.parameters[param] for param in common_params}
         for param in independent_params:
             params.update(
                 {
-                    f"{param}_{idx}": copy.deepcopy(inner.parameters[param])
+                    f"{param}_{idx}": param_like(model.parameters[param])
                     for idx in range(num_repetitions)
                 }
             )
 
-        super().__init__(parameters=params)
+        self.__model = model
+        self.__common_params = common_params
+        self.__independent_params = independent_params
+        self.__num_repetitions = num_repetitions
 
-        self.inner = inner
-        self.common_params = common_params
-        self.independent_params = independent_params
-        self.num_repetitions = num_repetitions
+        super().__init__(
+            parameters=params, internal_parameters=self.__model.internal_parameters
+        )
 
     def get_num_y_channels(self) -> int:
-        return self.num_repetitions * self.inner.get_num_y_channels()
+        return self.__num_repetitions * self.__model.get_num_y_channels()
 
     def can_rescale(self) -> Tuple[bool, bool]:
-        return self.inner.can_rescale()
+        return self.__model.can_rescale()
 
     def func(
         self,
         x: Array[("num_samples",), np.float64],
         param_values: Dict[str, float],
     ) -> Array[("num_y_channels", "num_samples"), np.float64]:
-        common_values = {param: param_values[param] for param in self.common_params}
+        values = {param: param_values[param] for param in self.__common_params}
 
-        ys = []
-        for idx in range(self.num_repetitions):
-            values = dict(common_values)
+        dim = self.__model.get_num_y_channels()
+        ys = np.zeros((self.get_num_y_channels(), len(x)))
+        for idx in range(self.__num_repetitions):
             values.update(
                 {
                     param: param_values[f"{param}_{idx}"]
-                    for param in self.independent_params
+                    for param in self.__independent_params
                 }
             )
-            ys.append(np.atleast_2d(self.inner.func(x, values)))
+            ys[idx * dim : (idx + 1) * dim, :] = np.atleast_2d(
+                self.__model.func(x, values)
+            )
 
-        return np.vstack(ys)
+        return ys
 
     def estimate_parameters(
         self,
         x: Array[("num_samples",), np.float64],
         y: Array[("num_y_channels", "num_samples"), np.float64],
     ):
-        dim = self.inner.get_num_y_channels()
+        dim = self.__model.get_num_y_channels()
 
-        common_params = {param: self.parameters[param] for param in self.common_params}
-        common_heuristics = {param: [] for param in self.common_params}
+        common_heuristics = {
+            param: [0.0] * self.__num_repetitions for param in self.__common_params
+        }
 
-        # FIXME - add "clear heuristics method instead"
-        inner_params = self.inner.parameters  # store for later
-
-        for idx in range(self.num_repetitions):
+        for idx in range(self.__num_repetitions):
             params = {
                 param: self.parameters[f"{param}_{idx}"]
-                for param in self.independent_params
+                for param in self.__independent_params
             }
 
-            params.update(common_params)
+            params.update(
+                {
+                    param: param_like(self.parameters[param])
+                    for param in self.__common_params
+                }
+            )
 
-            # Since common params are reused multiple times, clear their heuristic
-            # values between each use
-            for param_data in common_params.values():
-                param_data.heuristic = None
+            self.__model.parameters = params
+            self.__model.estimate_parameters(x, y[idx * dim : (idx + 1) * dim])
 
-            self.inner.parameters = params
-            self.inner.estimate_parameters(x, y[idx * dim : (idx + 1) * dim])
+            for param in self.__common_params:
+                common_heuristics[param][idx] = params[param].get_initial_value()
 
-            for param in self.common_params:
-                common_heuristics[param].append(params[param].get_initial_value())
-
-        # Combine the heuristics for the repetitions to find the best set of parameter
-        # values
-        for param in self.common_params:
+        # Combine the heuristics for the repetitions to find the best set of common
+        # parameter values
+        for param in self.__common_params:
             common_heuristics[param].append(np.mean(common_heuristics[param]))
 
         param_estimates = {
             param_name: self.parameters[param_name].get_initial_value()
             for param_name in self.parameters.keys()
-            if param_name not in self.common_params
+            if param_name not in self.__common_params
         }
-        costs = np.zeros(self.num_repetitions + 1)
-        for idx in range(self.num_repetitions + 1):
+        costs = np.zeros(self.__num_repetitions + 1)
+        for idx in range(self.__num_repetitions + 1):
             y_idx = self.__call__(
                 x=x,
                 **param_estimates,
                 **{
                     param_name: common_heuristics[param_name][idx]
-                    for param_name in self.common_params
+                    for param_name in self.__common_params
                 },
             )
             costs[idx] = np.sqrt(np.sum((y - y_idx) ** 2))
         best_heuristic = np.argmin(costs)
 
-        for param in self.common_params:
+        for param in self.__common_params:
             self.parameters[param].heuristic = common_heuristics[param][best_heuristic]
-
-        self.inner.parameters = inner_params
 
     def calculate_derived_params(
         self,
@@ -419,26 +443,34 @@ class RepeatedModel(Model):
         derived_params = {}
         derived_uncertainties = {}
 
-        dim = self.inner.get_num_y_channels()
+        dim = self.__model.get_num_y_channels()
 
-        for idx in range(self.num_repetitions):
+        # {param_name: [derived_values_from_each_repetition]}
+        derived_params_reps = None
+        derived_uncertainties_reps = None
+        derived_param_names = None
+
+        common_fitted_params = {
+            param: fitted_params[param] for param in self.__common_params
+        }
+        common_fit_uncertainties = {
+            param: fit_uncertainties[param] for param in self.__common_params
+        }
+
+        for idx in range(self.__num_repetitions):
             rep_params = {
                 param: fitted_params[f"{param}_{idx}"]
-                for param in self.independent_params
+                for param in self.__independent_params
             }
-            rep_params.update(
-                {param: fitted_params[param] for param in self.common_params}
-            )
+            rep_params.update(common_fitted_params)
 
             rep_uncertainties = {
                 param: fit_uncertainties[f"{param}_{idx}"]
-                for param in self.independent_params
+                for param in self.__independent_params
             }
-            rep_uncertainties.update(
-                {param: fit_uncertainties[param] for param in self.common_params}
-            )
+            rep_uncertainties.update(common_fit_uncertainties)
 
-            derived = self.inner.calculate_derived_params(
+            derived = self.__model.calculate_derived_params(
                 x=x,
                 y=y[idx * dim : (idx + 1) * dim],
                 fitted_params=rep_params,
@@ -446,48 +478,101 @@ class RepeatedModel(Model):
             )
             rep_derived_params, rep_derived_uncertainties = derived
 
-            derived_params.update(
-                {f"{param}_{idx}": value for param, value in rep_derived_params.items()}
-            )
-            derived_uncertainties.update(
-                {
-                    f"{param}_{idx}": value
-                    for param, value in rep_derived_uncertainties.items()
+            if derived_params_reps is None:
+                derived_param_names = rep_derived_params.keys()
+                # We can't preallocate prior to this because we don't know what the
+                # model's derived parameter values are until we've calculated them
+                derived_params_reps = {
+                    param_name: [0.0] * self.__num_repetitions
+                    for param_name in derived_param_names
                 }
-            )
+                derived_uncertainties_reps = {
+                    param_name: [0.0] * self.__num_repetitions
+                    for param_name in derived_param_names
+                }
 
-        # calculate statistical results
+            for param_name in derived_param_names:
+                derived_params_reps[param_name][idx] = rep_derived_params[param_name]
+                derived_uncertainties_reps[param_name][idx] = rep_derived_uncertainties[
+                    param_name
+                ]
+
         def add_statistics(values, uncertainties, param_names):
             for param_name in param_names:
                 param_values = np.array(
                     [
                         values[f"{param_name}_{idx}"]
-                        for idx in range(self.num_repetitions)
+                        for idx in range(self.__num_repetitions)
                     ]
                 )
                 param_uncerts = np.array(
                     [
                         uncertainties[f"{param_name}_{idx}"]
-                        for idx in range(self.num_repetitions)
+                        for idx in range(self.__num_repetitions)
                     ]
                 )
 
                 derived_params[f"{param_name}_mean"] = np.mean(param_values)
                 derived_uncertainties[f"{param_name}_mean"] = (
-                    np.sqrt(np.sum(param_uncerts**2)) / self.num_repetitions
+                    np.sqrt(np.sum(param_uncerts**2)) / self.__num_repetitions
                 )
 
                 derived_params[f"{param_name}_peak_peak"] = np.ptp(param_values)
                 derived_uncertainties[f"{param_name}_peak_peak"] = float("nan")
 
-        add_statistics(
-            derived_params, derived_uncertainties, param_names=rep_derived_params.keys()
-        )
-        add_statistics(
-            fitted_params,
-            fit_uncertainties,
-            param_names=self.independent_params,
-        )
+        if self.__aggregate_results:
+            keep = [
+                param_name
+                for param_name in derived_param_names
+                if np.all(
+                    np.isclose(
+                        derived_params_reps[param_name],
+                        derived_params_reps[param_name][0],
+                    )
+                )
+                and np.all(
+                    np.isclose(
+                        derived_uncertainties_reps[param_name],
+                        derived_uncertainties_reps[param_name][0],
+                    )
+                )
+            ]
+            derived_params.update(
+                {param_name: derived_params_reps[param_name][0] for param_name in keep}
+            )
+            derived_uncertainties.update(
+                {
+                    param_name: derived_uncertainties_reps[param_name][0]
+                    for param_name in keep
+                }
+            )
+        else:
+            for param_name in derived_param_names:
+                derived_params.update(
+                    {
+                        f"{param_name}_{idx}": derived_params_reps[param_name][idx]
+                        for idx in range(self.__num_repetitions)
+                    }
+                )
+                derived_uncertainties.update(
+                    {
+                        f"{param_name}_{idx}": derived_uncertainties_reps[param_name][
+                            idx
+                        ]
+                        for idx in range(self.__num_repetitions)
+                    }
+                )
+
+            add_statistics(
+                derived_params,
+                derived_uncertainties,
+                param_names=rep_derived_params.keys(),
+            )
+            add_statistics(
+                fitted_params,
+                fit_uncertainties,
+                param_names=self.__independent_params,
+            )
 
         return derived_params, derived_uncertainties